Implemented login page.
Everyone needs to reimport sql_files/user.sql because I added a test user.
Use test_guy@site.net and Abc123456 to login with a test user.
Given that the users Randy added have salt, password hash, and password fields I opted to use the PHP password_hash method. So new users won't have salt or password_sha256 fields in the database.

The login page now actually checks users, displays a failure message, or redirects to the home page.

The home page now displays if a user is logged in or not.