Reimplemented title query string
Can now use ?title= to set the title value on the search page.
However, I can't figure out how to set the range values in an acceptable way
so I have opted to not allow these to be set using a query string.

Also changed event listeners to 'change' instead of 'click' for filter inputs.